% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Pick a subset from the stats_picker class}
\usage{
pick(cls, var_levels)
}
\arguments{
\item{cls}{Object of the statspicker class.}

\item{var_levels}{List or vector containing quoted factor levels on which to
subset the data.}
}
