% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R, R/stats_picker.R
\name{pick.stats_picker}
\alias{pick.stats_picker}
\alias{stats_picker}
\title{Statistics picker}
\usage{
\method{pick}{stats_picker}(cls, var_levels)

stats_picker(df)
}
\section{Methods (by generic)}{
\itemize{
\item \code{pick}: Output a given crosstab
}}

