% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R, R/statspicker.R
\name{pick.statspicker}
\alias{pick.statspicker}
\alias{statspicker}
\title{Easily cross-tabulate categorical data}
\usage{
\method{pick}{statspicker}(cls, var_levels)

statspicker(df)
}
\arguments{
\item{df}{Target dataframe containing categorical data.}
}
\section{Methods (by generic)}{
\itemize{
\item \code{pick}: Output a given crosstab
}}

